+++ {{ cookiecutter.project_name }}/project/settings/production.py
@@ -9,3 +9,6 @@ class Production(ComposedProjectSettings):
     """
     # Labelize the deployed production environment name
     ENVIRONMENT = values.Value("Production", environ_name="DEPLOYED_ENVIRONMENT")
+
+    # Mark environement as indexable
+    SITE_INDEX_METAS = True